requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
praw>=7.7.1
requests>=2.31.0
streamlit>=1.26.0
